requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
backoff==2.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
distro==1.9.0
fastapi==0.112.0
ffmpy==0.4.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
gradio==4.41.0
gradio_client==1.3.0
h11==0.14.0
haystack-ai==2.3.1
haystack-experimental==0.1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
jiter==0.5.0
kiwisolver==1.4.5
lazy-imports==0.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
monotonic==1.6
more-itertools==10.4.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
ollama-haystack==0.0.7
openai==1.40.6
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
posthog==3.5.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
ruff==0.5.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tenacity==9.0.0
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
websockets==12.0
xxhash==3.4.1
yarl==1.9.4
